initial conditions now interpolate

+++ src/physics/fdtd/initial_conditions.hpp
@@ -15,7 +15,7 @@ using namespace kac_core::types;
 
 namespace kac_core::physics {
 
-	Matrix_1D raisedCosine1D(const int& size, const int& mu, const double& sigma) {
+	Matrix_1D raisedCosine1D(const int& size, const double& mu, const double& sigma) {
 		/*
 		Calculate a one dimensional raised cosine distribution.
 		See Bilbao, S. (2009) Numerical Sound Synthesis p.121.
@@ -41,7 +41,11 @@ namespace kac_core::physics {
 	}
 
 	Matrix_2D raisedCosine2D(
-		const int& size_X, const int& size_Y, const int& mu_x, const int& mu_y, const double& sigma
+		const int& size_X,
+		const int& size_Y,
+		const double& mu_x,
+		const double& mu_y,
+		const double& sigma
 	) {
 		/*
 		Calculate a two dimensional raised cosine distribution.
@@ -70,7 +74,8 @@ namespace kac_core::physics {
 		return raised_cosine;
 	}
 
-	Matrix_1D raisedTriangle1D(const int& size, const int& mu, const int& a, const int& b) {
+	Matrix_1D
+	raisedTriangle1D(const int& size, const double& mu, const double& a, const double& b) {
 		/*
 		Calculate a one dimensional triangular distribution.
 		See Bilbao, S. (2009) Numerical Sound Synthesis p.121.
